move lab service into view-data
