+++ containers/ecr-viewer/tests/unit/app/view-data/components/LabInfo.test.tsx
@@ -6,13 +6,13 @@ import { Bundle } from "fhir/r4";
 
 import BundleLab from "../../../../../../../test-data/fhir/BundleLab.json";
 import BundleLabNoLabIds from "../../../../../../../test-data/fhir/BundleLabNoLabIds.json";
-import {
-  evaluateLabInfoData,
-  LabReportElementData,
-} from "@/app/services/labsService";
 import { evaluateAll } from "@/app/utils/evaluate";
 import fhirPathMappings from "@/app/utils/evaluate/fhir-paths";
 import LabInfo from "@/app/view-data/components/LabInfo";
+import {
+  evaluateLabInfoData,
+  LabReportElementData,
+} from "@/app/view-data/services/labsService";
 
 describe("LabInfo", () => {
   describe("when labResults is LabReportElementData[]", () => {
